+++ packages/br-validators/src/strip/index.ts
@@ -3,4 +3,5 @@ export { stripCpf } from './cpf.js';
 export { stripCnpj } from './cnpj.js';
 export { stripCnh } from './cnh.js';
 export { stripRenavam } from './renavam.js';
+export { stripTituloEleitor } from './titulo-eleitor.js';
 export { stripPlaca } from './placa.js';